style: separate BuildAssetCache from Download in a new file
